[FLINK-6177] Combine the streaming & batch distributed cache tests and inherit from StreamingMultipleProgramsTestBase
